Refactor app/home and routes
