+++ src/containers/App.js
@@ -1,0 +1,14 @@
+import React, {createClass} from 'react'
+import {RouteHandler} from 'react-router'
+
+import Home from '../pages/Home'
+
+let App = createClass({
+  render() {
+    return (
+      <div><Home /></div>
+    )
+  }
+})
+
+export default App
